Last Commit First Week

+++ src/main/java/com/example/testweb/BookRepository.java
@@ -2,5 +2,7 @@ package com.example.testweb;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer>
+{
+
 }
